new route for hide-keeper-on-site action and moving it to ExtPreferenceController
